+++ tests/test_admin.py
@@ -1,5 +1,6 @@
 import json
 import os
+from unittest import mock
 
 import django
 import django.core.files
@@ -411,14 +412,17 @@ class FilerImageAdminUrlsTests(TestCase):
             Image.objects.create(owner=self.superuser, original_filename='some-image.jpg'),  # missing file
             Image.objects.create(owner=self.superuser, original_filename='some-image.jpg', file=self.file_object.file),
             Image.objects.create(owner=self.superuser, original_filename='some-image.svg', file=self.file_object.file),
+            # SVGs are not necessarily stored as Image, see #1590
+            File.objects.create(owner=self.superuser, original_filename='some-file.svg', file=self.file_object.file),
         ]
         test_set = [
-            (files[0], 'text/plain', None),
-            (files[1], 'image/jpeg', None),
-            (files[2], 'image/jpeg', files[2].file.url),
-            (files[3], 'image/svg+xml', reverse(admin_urlname(Image._meta, 'expand'), args=(files[3].pk,))),
+            (files[0], 'text/plain', False),
+            (files[1], 'image/jpeg', False),
+            (files[2], 'image/jpeg', True),
+            (files[3], 'image/svg+xml', True),
+            (files[4], 'image/svg+xml', True),
         ]
-        for file, mime_type, expected_url in test_set:
+        for file, mime_type, has_expand_link in test_set:
             file.mime_type = mime_type
             file.save()
             models = [File]
@@ -427,10 +431,15 @@ class FilerImageAdminUrlsTests(TestCase):
             for model in models:
                 response = self.client.get(reverse(admin_urlname(model._meta, 'change'),
                                                    kwargs={'object_id': file.pk}))
-                if expected_url:
+                if not has_expand_link:
+                    self.assertNotContains(response, 'filer-icon-expand')
+                elif 'svg' in mime_type:
+                    # Each admin links to its own expand view, so that the file can be
+                    # looked up by the model the admin is registered for
+                    expected_url = reverse(admin_urlname(model._meta, 'expand'), args=(file.pk,))
                     self.assertContains(response, f'href="{expected_url}"')
                 else:
-                    self.assertNotContains(response, 'filer-icon-expand')
+                    self.assertContains(response, f'href="{file.file.url}"')
 
     def test_image_expand_view(self):
         url = reverse(admin_urlname(Image._meta, 'expand'), kwargs={
@@ -444,6 +453,62 @@ class FilerImageAdminUrlsTests(TestCase):
         self.assertContains(
             response,
             f"""<img id="img" src="{original_url}" onclick="this.classList.toggle('zoom')"/>"""
+        )
+
+    def test_file_expand_view(self):
+        """SVGs stored as File (not Image) have an expand view, too, see #1590"""
+        file = File.objects.create(
+            owner=self.superuser,
+            original_filename='some-file.svg',
+            file=self.file_object.file,
+            mime_type='image/svg+xml',
+        )
+        url = reverse(admin_urlname(File._meta, 'expand'), kwargs={'file_id': file.pk})
+
+        response = self.client.get(url)
+
+        self.assertEqual(url, file.get_admin_expand_view_url())
+        self.assertContains(
+            response,
+            f"""<img id="img" src="{file.url}" onclick="this.classList.toggle('zoom')"/>"""
+        )
+
+    def test_expand_view_requires_read_permission(self):
+        """The expand view does not disclose files the user may not read"""
+        file = File.objects.create(
+            owner=self.superuser,
+            original_filename='some-file.svg',
+            file=self.file_object.file,
+            mime_type='image/svg+xml',
+            folder=Folder.objects.create(name='svgs', owner=self.superuser),
+        )
+        url = reverse(admin_urlname(File._meta, 'expand'), kwargs={'file_id': file.pk})
+
+        with mock.patch.object(File, 'has_read_permission', return_value=False):
+            response = self.client.get(url)
+
+        self.assertEqual(response.status_code, 404)
+
+    def test_directory_listing_with_svg_file(self):
+        """A folder containing an SVG stored as File can be listed, see #1590"""
+        folder = Folder.objects.create(name='svgs', owner=self.superuser)
+        file = File.objects.create(
+            owner=self.superuser,
+            original_filename='some-file.svg',
+            file=self.file_object.file,
+            folder=folder,
+        )
+        # The MIME type is what counts, it is reset when the file contents change
+        file.mime_type = 'image/svg+xml'
+        file.save()
+
+        response = self.client.get(reverse('admin:filer-directory_listing',
+                                           kwargs={'folder_id': folder.pk}))
+
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(
+            response,
+            reverse(admin_urlname(File._meta, 'expand'), args=(file.pk,)),
         )